requirements.txt
streamlit
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
PyPDF2
langchain
faiss-cpu
sentence-transformers
langchain-community
tavily-python